コンポーネント定義をまとめる (循環インポートを避けるため)
