--- ptcs/ptcs_control/components/__init__.py
@@ -1,12 +1,0 @@
-from .direction import Direction
-from .joint import Joint
-from .junction import Junction
-from .section import Section
-
-
-__all__ = [
-    "Direction",
-    "Joint",
-    "Junction",
-    "Section",
-]
