--- ptcs/ptcs_control/components/direction.py
@@ -1,17 +1,0 @@
-from enum import Enum, auto
-
-
-class Direction(Enum):
-    """
-    サーボモーターの方向
-
-    ```
-    _______________
-    ______  _______ STRAIGHT
-          \ \______
-           \_______ CURVE
-    ```
-    """
-
-    STRAIGHT = auto()
-    CURVE = auto()
