+++ ptcs/ptcs_control/components/length.py
@@ -1,0 +1,2 @@
+STRAIGHT_UNIT = 215.0  # 直線レールの長さ [mm]
+CURVE_UNIT = 169.0  # 曲線レールの長さ [mm]
